+++ scripts/call_consensus.py
@@ -40,6 +40,12 @@ if __name__ == '__main__':
                                 forward and reverse strands to make a call. The numerator of this fraction is the number of high-quality
                                 consensus-supporting reads on one strand.  The denominator is the total number of high-quality consensus-supporting 
                                 reads on both strands combined."""
+    help['vcfFileName']    = """VCF Output file name. If specified, a VCF file with this file name will be created in the same directory as the
+                                consensus fasta file for this sample."""
+    help['vcfRefName']     = """Name of the reference file.  This is only used in the generated VCF file header."""
+    help['vcfAllPos']      = """Flag to cause VCF file generation at all positions, not just the snp positions.  This has no effect on
+                                the consensus fasta file, it only affects the VCF file.  This capability is intended primarily as a diagnostic tool and
+                                enabling this flag will greatly increase execution time."""
     help['verbose']        = """Verbose message level (0=no info, 5=lots)"""
 
     parser.add_argument(                          dest='allPileupFile',  type=str,                                                        help=help['allPileupFile'])
@@ -50,6 +56,9 @@ if __name__ == '__main__':
     parser.add_argument('-c', '--minConsFreq',    dest='minConsFreq',    type=minConsFreq,    default=0.60,               metavar='FREQ', help=help['minConsFreq'])
     parser.add_argument('-d', '--minConsStrdDpth',dest='minConsStrdDpth',type=int,            default=0,                  metavar='INT',  help=help['minConsStrdDpth'])
     parser.add_argument('-b', '--minConsStrdBias',dest='minConsStrdBias',type=minConsStrdBias,default=0,                  metavar='FREQ', help=help['minConsStrdBias'])
+    parser.add_argument(      '--vcfFileName',    dest='vcfFileName',    type=str,            default=None,               metavar='NAME', help=help['vcfFileName'])
+    parser.add_argument(      '--vcfRefName',     dest='vcfRefName',     type=str,            default='Unknown reference',metavar='NAME', help=help['vcfRefName'])
+    parser.add_argument(      '--vcfAllPos',      dest='vcfAllPos',      action='store_true',                                             help=help['vcfAllPos'])
     parser.add_argument('-v', '--verbose',        dest='verbose',        type=int,            default=1,                  metavar='0..5', help=help['verbose'])
     parser.add_argument('--version', action='version', version='%(prog)s version ' + __version__)
     args_dict = vars(parser.parse_args())